docs: refactor more code to be re-usable.
